Fix exception thrown when nonexistent filename, especially `-` for stdin, is passed
The error is caused by Snapshot class naively reading in file content to
create a md5 digest for the file. This is not possible when `-` is passed,
because a) it is most likely not an existing filename, and b) even if we
were to interpret it as stdin, it isn't seekable
(outside of using `os.lseek` when a real file was redirected).

The fix: when calculating digests for a non-existent file, just invent
one from current time and a random value. It will be a new value every
time, ensuring no caching between such stdin invocations.